+++ CHANGELOG.md
@@ -7,8 +7,9 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - `operator-rs` `0.4.0` → `0.7.0` ([#81]).
-- TLS certificates now provided via config map instead of custom resource ([#81]).
 - User authentication now provided via secret instead of custom resource ([#81]).
+- User authentication not exposed in configmap anymore ([#81]).
+- TLS certificates / keystore now retrieved via secret-operator ([#81]).
 - The Trino version is now string instead of enum ([#81]).
 
 [#81]: https://github.com/stackabletech/trino-operator/pull/81